Add PostgreSQL support and update model definitions
Added PostgreSQL connectivity in settings and introduced a build script for Django project. The model definitions were updated with changes to primary key handling and a progress bar was added for fixtures generation. Additionally, the Django's static file storage is set to use the WhiteNoise storage backend for improved performance with compressed files.