Add creation of default superuser after migration
During the ready state in accounts/app.py, a function has been added to create a default superuser if it does not already exist, alongside the existing function that creates a default site. This is done upon migration, using the post_migrate signal. The superuser has a fixed username and password, and is associated with site 0.