Update logging configuration in local and production settings
The configuration of the logging system in both local and production settings has been updated. This change ensures that the logging system outputs to console for both custom applications and Django-related logs. Also, a new handler for console output has been added in the production settings which was previously writing logs to a file.