Integrate django-environ for environment settings
This update replaces the previous way of extracting environment variables in wsgi.py and manage.py by using django-environ. The changes allow for automatic reading of a .env file if the 'DJANGO_SETTINGS_READ_ENV_FILE' variable is set to True. This aims to make our app more convenient to configure according to various environments.